自島対象のPlanをOwnIsland内に移動
