+++ app/app/Entity/Plan/ForeignIsland/FundsTransportToForeignIslandPlan.php
@@ -8,7 +8,7 @@ use App\Entity\Cell\Others\Shallow;
 use App\Entity\Cell\Ship\TransportShip;
 use App\Entity\Log\LogRow\FundsTransportationLog;
 use App\Entity\Log\Logs;
-use App\Entity\Plan\FundsTransportationPlan;
+use App\Entity\Plan\OwnIsland\FundsTransportationPlan;
 use App\Entity\Status\Status;
 use App\Entity\Terrain\Terrain;
 use App\Models\Island;
