+++ app/app/Entity/Plan/OwnIsland/AfforestationPlan.php
@@ -1,25 +1,27 @@
 <?php
 
-namespace App\Entity\Plan;
+namespace App\Entity\Plan\OwnIsland;
 
-use App\Entity\Cell\Others\Volcano;
-use App\Entity\Cell\ResourcesProduction\Mine;
+use App\Entity\Cell\HasWoods\Forest;
+use App\Entity\Log\LogConst;
 use App\Entity\Log\LogRow\AbortInvalidCellLog;
 use App\Entity\Log\LogRow\AbortLackOfFundsLog;
+use App\Entity\Log\LogRow\AfforestationLog;
 use App\Entity\Log\LogRow\ExecuteLog;
 use App\Entity\Log\Logs;
+use App\Entity\Plan\Plan;
 use App\Entity\Status\Status;
 use App\Entity\Terrain\Terrain;
 use App\Models\Island;
 use App\Models\Turn;
 use Illuminate\Support\Collection;
 
-class ConstructMinePlan extends Plan
+class AfforestationPlan extends Plan
 {
-    public const KEY = 'construct_mine';
+    public const KEY = 'afforestation';
 
-    public const NAME = '採掘場整備';
-    public const PRICE = 300;
+    public const NAME = '植林';
+    public const PRICE = 50;
     public const PRICE_STRING = '(' . self::PRICE . '億円)';
 
     protected string $key = self::KEY;
@@ -29,19 +31,21 @@ class ConstructMinePlan extends Plan
     public function execute(Island $island, Terrain $terrain, Status $status, Turn $turn, Collection $foreignIslandTargetedPlans): ExecutePlanResult
     {
         $cell = $terrain->getCell($this->point);
+        $logs = Logs::create();
         if ($status->getFunds() < self::PRICE) {
-            $logs = Logs::create()->add(new AbortLackOfFundsLog($island, $this));
+            $logs->add(new AbortLackOfFundsLog($island, $this));
             return new ExecutePlanResult($terrain, $status, $logs, false);
         }
 
-        if (!in_array($cell::TYPE, [Volcano::TYPE], true)) {
-            $logs = Logs::create()->add(new AbortInvalidCellLog($island, $this, $cell));
+        if (!in_array($cell::TYPE, self::CONSTRUCTABLE_CELLS, true)) {
+            $logs->add(new AbortInvalidCellLog($island, $this, $cell));
             return new ExecutePlanResult($terrain, $status, $logs, false);
         }
 
-        $terrain->setCell($this->point, new Mine(point: $this->point));
+        $terrain->setCell($this->point, new Forest(point: $this->point));
         $status->setFunds($status->getFunds() - self::PRICE);
-        $logs = Logs::create()->add(new ExecuteLog($island, $this));
+        $logs->add(new ExecuteLog($island, $this, LogConst::VISIBILITY_PRIVATE));
+        $logs->add(new AfforestationLog($island));
         return new ExecutePlanResult($terrain, $status, $logs, true);
     }
 }
