+++ app/app/Entity/Plan/OwnIsland/AttractActivitiesPlan.php
@@ -1,30 +1,26 @@
 <?php
 
-namespace App\Entity\Plan;
+namespace App\Entity\Plan\OwnIsland;
 
-use App\Entity\Cell\Cell;
-use App\Entity\Cell\Others\Sea;
-use App\Entity\Cell\Others\Shallow;
-use App\Entity\Cell\Ship\Battleship;
-use App\Entity\Log\LogRow\AbortInvalidTerrainLog;
+use App\Entity\Cell\HasPopulation\IHasPopulation;
 use App\Entity\Log\LogRow\AbortLackOfFundsLog;
 use App\Entity\Log\LogRow\ExecuteLog;
 use App\Entity\Log\Logs;
+use App\Entity\Plan\Plan;
 use App\Entity\Status\Status;
 use App\Entity\Terrain\Terrain;
 use App\Models\Island;
 use App\Models\Turn;
 use Illuminate\Support\Collection;
 
-class ConstructBattleshipPlan extends Plan
+class AttractActivitiesPlan extends Plan
 {
-    public const KEY = 'construct_battleship';
+    public const KEY = 'attract_activities';
 
-    public const NAME = '戦艦建造';
-    public const PRICE = 500;
+    public const NAME = '誘致活動';
+    public const PRICE = 1000;
     public const PRICE_STRING = '(' . self::PRICE . '億円)';
     public const AMOUNT_STRING = '(:amount:回実施)';
-    public const DEFAULT_AMOUNT_STRING = '(1回実施)';
     public const USE_POINT = false;
     public const USE_AMOUNT = true;
 
@@ -34,35 +30,27 @@ class ConstructBattleshipPlan extends Plan
     protected bool $usePoint = self::USE_POINT;
     protected bool $useAmount = self::USE_AMOUNT;
     protected string $amountString = self::AMOUNT_STRING;
-    protected string $defaultAmountString = self::DEFAULT_AMOUNT_STRING;
 
     public function execute(Island $island, Terrain $terrain, Status $status, Turn $turn, Collection $foreignIslandTargetedPlans): ExecutePlanResult
     {
+        $logs = Logs::create();
+
         if ($status->getFunds() < self::PRICE) {
             $this->amount = 0;
-            $logs = Logs::create()->add(new AbortLackOfFundsLog($island, $this));
+            $logs->add(new AbortLackOfFundsLog($island, $this));
             return new ExecutePlanResult($terrain, $status, $logs, false);
         }
 
-        $seaCells = $terrain->findByTypes([Sea::TYPE, Shallow::TYPE]);
+        /** @var IHasPopulation $cell */
+        $hasPopulationCells = $terrain->findByInterface(IHasPopulation::class);
 
-        if ($seaCells->count() <= 0) {
-            $this->amount = 0;
-            $logs = Logs::create()->add(new AbortInvalidTerrainLog($island, $this));
-            return new ExecutePlanResult($terrain, $status, $logs, false);
+        foreach ($hasPopulationCells as $cell) {
+            $addPopulation = random_int(1, 3) * 100;
+            $cell->setPopulation($cell->getPopulation() + $addPopulation);
         }
 
         $this->amount -= 1;
 
-        /** @var Cell $cell */
-        $cell = $seaCells->random();
-
-        $terrain->setCell($cell->getPoint(), new Battleship(
-            point: $cell->getPoint(),
-            elevation: $cell->getElevation(),
-            affiliation_id: $island->id,
-            affiliation_name: $island->name,
-        ));
         $status->setFunds($status->getFunds() - self::PRICE);
         $logs = Logs::create()->add(new ExecuteLog($island, $this));
         return new ExecutePlanResult($terrain, $status, $logs, true);
