+++ app/app/Entity/Plan/OwnIsland/ConstructFactoryPlan.php
@@ -1,25 +1,25 @@
 <?php
 
-namespace App\Entity\Plan;
+namespace App\Entity\Plan\OwnIsland;
 
-use App\Entity\Cell\Others\Shallow;
-use App\Entity\Cell\ResourcesProduction\Oilfield;
+use App\Entity\Cell\FundsProduction\Factory;
 use App\Entity\Log\LogRow\AbortInvalidCellLog;
 use App\Entity\Log\LogRow\AbortLackOfFundsLog;
 use App\Entity\Log\LogRow\ExecuteLog;
 use App\Entity\Log\Logs;
+use App\Entity\Plan\Plan;
 use App\Entity\Status\Status;
 use App\Entity\Terrain\Terrain;
 use App\Models\Island;
 use App\Models\Turn;
 use Illuminate\Support\Collection;
 
-class ConstructOilfieldPlan extends Plan
+class ConstructFactoryPlan extends Plan
 {
-    public const KEY = 'construct_oilfield';
+    public const KEY = 'construct_factory';
 
-    public const NAME = '油田発掘';
-    public const PRICE = 200;
+    public const NAME = '工場建設';
+    public const PRICE = 20;
     public const PRICE_STRING = '(' . self::PRICE . '億円)';
 
     protected string $key = self::KEY;
@@ -34,12 +34,12 @@ class ConstructOilfieldPlan extends Plan
             return new ExecutePlanResult($terrain, $status, $logs, false);
         }
 
-        if (!in_array($cell::TYPE, [Shallow::TYPE], true)) {
+        if (!in_array($cell::TYPE, self::CONSTRUCTABLE_CELLS, true)) {
             $logs = Logs::create()->add(new AbortInvalidCellLog($island, $this, $cell));
             return new ExecutePlanResult($terrain, $status, $logs, false);
         }
 
-        $terrain->setCell($this->point, new Oilfield(point: $this->point));
+        $terrain->setCell($this->point, new Factory(point: $this->point));
         $status->setFunds($status->getFunds() - self::PRICE);
         $logs = Logs::create()->add(new ExecuteLog($island, $this));
         return new ExecutePlanResult($terrain, $status, $logs, true);
