+++ app/app/Entity/Plan/OwnIsland/ConstructFarmDomePlan.php
@@ -1,13 +1,15 @@
 <?php
 
-namespace App\Entity\Plan;
+namespace App\Entity\Plan\OwnIsland;
 
-use App\Entity\Cell\Others\Plain;
+use App\Entity\Cell\FoodsProduction\Farm;
+use App\Entity\Cell\FoodsProduction\FarmDome;
 use App\Entity\Log\LogRow\AbortInvalidCellLog;
 use App\Entity\Log\LogRow\AbortLackOfFundsLog;
 use App\Entity\Log\LogRow\AbortNoDevelopmentPointsLog;
 use App\Entity\Log\LogRow\ExecuteLog;
 use App\Entity\Log\Logs;
+use App\Entity\Plan\Plan;
 use App\Entity\Status\DevelopmentPointsConst;
 use App\Entity\Status\Status;
 use App\Entity\Terrain\Terrain;
@@ -15,19 +17,19 @@ use App\Models\Island;
 use App\Models\Turn;
 use Illuminate\Support\Collection;
 
-class GroundLevelingPlan extends Plan
+class ConstructFarmDomePlan extends Plan
 {
-    public const KEY = 'ground_leveling';
+    public const KEY = 'construct_farm_dome';
 
-    public const NAME = '高速整地';
-    public const PRICE = 100;
+    public const NAME = '農場ドーム化';
+    public const PRICE = 2000;
     public const PRICE_STRING = '(' . self::PRICE . '億円)';
-    public const EXECUTABLE_DEVELOPMENT_POINT = DevelopmentPointsConst::GROUND_LEVELING_AVAILABLE_POINTS;
+
+    public const EXECUTABLE_DEVELOPMENT_POINT = DevelopmentPointsConst::CONSTRUCT_FARM_DOME_AVAILABLE_POINTS;
 
     protected string $key = self::KEY;
     protected string $name = self::NAME;
     protected int $price = self::PRICE;
-    protected int $executableDevelopmentPoint = self::EXECUTABLE_DEVELOPMENT_POINT;
 
     public function execute(Island $island, Terrain $terrain, Status $status, Turn $turn, Collection $foreignIslandTargetedPlans): ExecutePlanResult
     {
@@ -44,14 +46,14 @@ class GroundLevelingPlan extends Plan
             return new ExecutePlanResult($terrain, $status, $logs, false);
         }
 
-        if (!in_array($cell::TYPE, self::GRADABLE_CELLS, true)) {
+        if ($cell::TYPE !== Farm::TYPE) {
             $logs->add(new AbortInvalidCellLog($island, $this, $cell));
             return new ExecutePlanResult($terrain, $status, $logs, false);
         }
 
-        $terrain->setCell($this->point, new Plain(point: $this->point));
+        $terrain->setCell($this->point, new FarmDome(point: $this->point));
         $status->setFunds($status->getFunds() - self::PRICE);
         $logs->add(new ExecuteLog($island, $this));
-        return new ExecutePlanResult($terrain, $status, $logs, false);
+        return new ExecutePlanResult($terrain, $status, $logs, true);
     }
 }
