+++ app/app/Entity/Plan/OwnIsland/ConstructMissileBasePlan.php
@@ -1,24 +1,27 @@
 <?php
 
-namespace App\Entity\Plan;
+namespace App\Entity\Plan\OwnIsland;
 
-use App\Entity\Cell\FundsProduction\Factory;
+use App\Entity\Cell\MissileFireable\MissileBase;
+use App\Entity\Log\LogConst;
 use App\Entity\Log\LogRow\AbortInvalidCellLog;
 use App\Entity\Log\LogRow\AbortLackOfFundsLog;
+use App\Entity\Log\LogRow\AfforestationLog;
 use App\Entity\Log\LogRow\ExecuteLog;
 use App\Entity\Log\Logs;
+use App\Entity\Plan\Plan;
 use App\Entity\Status\Status;
 use App\Entity\Terrain\Terrain;
 use App\Models\Island;
 use App\Models\Turn;
 use Illuminate\Support\Collection;
 
-class ConstructFactoryPlan extends Plan
+class ConstructMissileBasePlan extends Plan
 {
-    public const KEY = 'construct_factory';
+    public const KEY = 'construct_missile_base';
 
-    public const NAME = '工場建設';
-    public const PRICE = 20;
+    public const NAME = 'ミサイル基地建設';
+    public const PRICE = 300;
     public const PRICE_STRING = '(' . self::PRICE . '億円)';
 
     protected string $key = self::KEY;
@@ -28,19 +31,21 @@ class ConstructFactoryPlan extends Plan
     public function execute(Island $island, Terrain $terrain, Status $status, Turn $turn, Collection $foreignIslandTargetedPlans): ExecutePlanResult
     {
         $cell = $terrain->getCell($this->point);
+        $logs = Logs::create();
         if ($status->getFunds() < self::PRICE) {
-            $logs = Logs::create()->add(new AbortLackOfFundsLog($island, $this));
+            $logs->add(new AbortLackOfFundsLog($island, $this));
             return new ExecutePlanResult($terrain, $status, $logs, false);
         }
 
         if (!in_array($cell::TYPE, self::CONSTRUCTABLE_CELLS, true)) {
-            $logs = Logs::create()->add(new AbortInvalidCellLog($island, $this, $cell));
+            $logs->add(new AbortInvalidCellLog($island, $this, $cell));
             return new ExecutePlanResult($terrain, $status, $logs, false);
         }
 
-        $terrain->setCell($this->point, new Factory(point: $this->point));
+        $terrain->setCell($this->point, new MissileBase(point: $this->point));
         $status->setFunds($status->getFunds() - self::PRICE);
-        $logs = Logs::create()->add(new ExecuteLog($island, $this));
+        $logs->add(new ExecuteLog($island, $this, LogConst::VISIBILITY_PRIVATE));
+        $logs->add(new AfforestationLog($island));
         return new ExecutePlanResult($terrain, $status, $logs, true);
     }
 }
