+++ app/app/Entity/Plan/OwnIsland/ConstructOilfieldPlan.php
@@ -1,26 +1,26 @@
 <?php
 
-namespace App\Entity\Plan;
+namespace App\Entity\Plan\OwnIsland;
 
-use App\Entity\Cell\MissileFireable\MissileBase;
+use App\Entity\Cell\Others\Shallow;
+use App\Entity\Cell\ResourcesProduction\Oilfield;
 use App\Entity\Log\LogRow\AbortInvalidCellLog;
 use App\Entity\Log\LogRow\AbortLackOfFundsLog;
-use App\Entity\Log\LogRow\AfforestationLog;
 use App\Entity\Log\LogRow\ExecuteLog;
 use App\Entity\Log\Logs;
-use App\Entity\Log\LogConst;
+use App\Entity\Plan\Plan;
 use App\Entity\Status\Status;
 use App\Entity\Terrain\Terrain;
 use App\Models\Island;
 use App\Models\Turn;
 use Illuminate\Support\Collection;
 
-class ConstructMissileBasePlan extends Plan
+class ConstructOilfieldPlan extends Plan
 {
-    public const KEY = 'construct_missile_base';
+    public const KEY = 'construct_oilfield';
 
-    public const NAME = 'ミサイル基地建設';
-    public const PRICE = 300;
+    public const NAME = '油田発掘';
+    public const PRICE = 200;
     public const PRICE_STRING = '(' . self::PRICE . '億円)';
 
     protected string $key = self::KEY;
@@ -30,21 +30,19 @@ class ConstructMissileBasePlan extends Plan
     public function execute(Island $island, Terrain $terrain, Status $status, Turn $turn, Collection $foreignIslandTargetedPlans): ExecutePlanResult
     {
         $cell = $terrain->getCell($this->point);
-        $logs = Logs::create();
         if ($status->getFunds() < self::PRICE) {
-            $logs->add(new AbortLackOfFundsLog($island, $this));
+            $logs = Logs::create()->add(new AbortLackOfFundsLog($island, $this));
             return new ExecutePlanResult($terrain, $status, $logs, false);
         }
 
-        if (!in_array($cell::TYPE, self::CONSTRUCTABLE_CELLS, true)) {
-            $logs->add(new AbortInvalidCellLog($island, $this, $cell));
+        if (!in_array($cell::TYPE, [Shallow::TYPE], true)) {
+            $logs = Logs::create()->add(new AbortInvalidCellLog($island, $this, $cell));
             return new ExecutePlanResult($terrain, $status, $logs, false);
         }
 
-        $terrain->setCell($this->point, new MissileBase(point: $this->point));
+        $terrain->setCell($this->point, new Oilfield(point: $this->point));
         $status->setFunds($status->getFunds() - self::PRICE);
-        $logs->add(new ExecuteLog($island, $this, LogConst::VISIBILITY_PRIVATE));
-        $logs->add(new AfforestationLog($island));
+        $logs = Logs::create()->add(new ExecuteLog($island, $this));
         return new ExecutePlanResult($terrain, $status, $logs, true);
     }
 }
