+++ app/app/Entity/Plan/OwnIsland/ExcavationPlan.php
@@ -1,19 +1,18 @@
 <?php
 
-namespace App\Entity\Plan;
+namespace App\Entity\Plan\OwnIsland;
 
-use App\Entity\Cell\Cell;
-use App\Entity\Cell\CellConst;
-use App\Entity\Cell\Others\Lake;
+use App\Entity\Cell\Others\Mountain;
 use App\Entity\Cell\Others\Sea;
 use App\Entity\Cell\Others\Shallow;
+use App\Entity\Cell\Others\Volcano;
 use App\Entity\Cell\Others\Wasteland;
 use App\Entity\Log\LogRow\AbortInvalidCellLog;
 use App\Entity\Log\LogRow\AbortLackOfFundsLog;
-use App\Entity\Log\LogRow\AbortNoAdjacentLandsLog;
 use App\Entity\Log\LogRow\AbortNoDevelopmentPointsLog;
 use App\Entity\Log\LogRow\ExecuteLog;
 use App\Entity\Log\Logs;
+use App\Entity\Plan\Plan;
 use App\Entity\Status\DevelopmentPointsConst;
 use App\Entity\Status\Status;
 use App\Entity\Terrain\Terrain;
@@ -21,14 +20,15 @@ use App\Models\Island;
 use App\Models\Turn;
 use Illuminate\Support\Collection;
 
-class LandfillPlan extends Plan
+class ExcavationPlan extends Plan
 {
-    public const KEY = 'landfill';
+    public const KEY = 'excavation';
 
-    public const NAME = '埋め立て';
-    public const PRICE = 150;
+    public const NAME = '掘削';
+    public const PRICE = 200;
     public const PRICE_STRING = '(' . self::PRICE . '億円)';
-    public const EXECUTABLE_DEVELOPMENT_POINT = DevelopmentPointsConst::LANDFILL_AVAILABLE_POINTS;
+
+    public const EXECUTABLE_DEVELOPMENT_POINT = DevelopmentPointsConst::EXCAVATION_AVAILABLE_POINTS;
 
     protected string $key = self::KEY;
     protected string $name = self::NAME;
@@ -39,7 +39,6 @@ class LandfillPlan extends Plan
     {
         $cell = $terrain->getCell($this->point);
         $logs = Logs::create();
-
         if ($status->getFunds() < self::PRICE) {
             $logs->add(new AbortLackOfFundsLog($island, $this));
             return new ExecutePlanResult($terrain, $status, $logs, false);
@@ -50,39 +49,15 @@ class LandfillPlan extends Plan
             return new ExecutePlanResult($terrain, $status, $logs, false);
         }
 
-        if (!in_array($cell::TYPE, [Shallow::TYPE, Sea::TYPE, Lake::TYPE], true)) {
+        if (!in_array($cell::TYPE, array_merge([Shallow::TYPE, Mountain::TYPE, Volcano::TYPE], self::CONSTRUCTABLE_CELLS), true)) {
             $logs->add(new AbortInvalidCellLog($island, $this, $cell));
             return new ExecutePlanResult($terrain, $status, $logs, false);
         }
 
-        $landCells = $terrain->getAroundCells($cell->getPoint())->filter(function ($cell) {
-            return $cell::ATTRIBUTE[CellConst::IS_LAND];
-        });
-
-        if ($landCells->count() === 0) {
-            $logs->add(new AbortNoAdjacentLandsLog($island, $this));
-            return new ExecutePlanResult($terrain, $status, $logs, false);
-        }
-
-        if (in_array($cell::TYPE, [Shallow::TYPE, Lake::TYPE], true)) {
+        if ($cell::TYPE === Shallow::TYPE) {
+            $terrain->setCell($this->point, new Sea(point: $this->point));
+        } else if (in_array($cell::TYPE, [Mountain::TYPE, Volcano::TYPE], true)) {
             $terrain->setCell($this->point, new Wasteland(point: $this->point));
-
-            // 周囲が3セル以上陸地だった場合、周囲の海は浅瀬になる
-            $cells = $terrain->getAroundCells($this->point);
-            $aroundGroundCount = 0;
-            /** @var Cell $c */
-            foreach ($cells as $c) {
-                if ($c::ATTRIBUTE[CellConst::IS_LAND]) {
-                    $aroundGroundCount += 1;
-                }
-            }
-            if ($aroundGroundCount >= 3) {
-                foreach ($cells as $c) {
-                    if ($c::TYPE === Sea::TYPE) {
-                        $terrain->setCell($c->getPoint(), new Shallow(point: $c->getPoint()));
-                    }
-                }
-            }
         } else {
             $terrain->setCell($this->point, new Shallow(point: $this->point));
         }
