+++ app/app/Entity/Plan/OwnIsland/FiringMissilePlan.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App\Entity\Plan;
+namespace App\Entity\Plan\OwnIsland;
 
 use App\Entity\Cell\Cell;
 use App\Entity\Cell\CellConst;
@@ -19,6 +19,7 @@ use App\Entity\Log\LogRow\MissileSelfDestructLog;
 use App\Entity\Log\LogRow\SoldMonsterCorpseLog;
 use App\Entity\Log\Logs;
 use App\Entity\Plan\ForeignIsland\FiringMissileToForeignIslandPlan;
+use App\Entity\Plan\Plan;
 use App\Entity\Status\Status;
 use App\Entity\Terrain\Terrain;
 use App\Models\Island;
