+++ app/app/Entity/Plan/OwnIsland/GradingPlan.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App\Entity\Plan;
+namespace App\Entity\Plan\OwnIsland;
 
 use App\Entity\Cell\Others\Plain;
 use App\Entity\Log\LogRow\AbortInvalidCellLog;
@@ -8,6 +8,7 @@ use App\Entity\Log\LogRow\AbortLackOfFundsLog;
 use App\Entity\Log\LogRow\ExecuteLog;
 use App\Entity\Log\LogRow\FindBuriedTreasureLog;
 use App\Entity\Log\Logs;
+use App\Entity\Plan\Plan;
 use App\Entity\Status\Status;
 use App\Entity\Terrain\Terrain;
 use App\Entity\Util\Rand;
